% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperUtils.R
\name{addfiglab}
\alias{addfiglab}
\alias{.addfiglab}
\title{Add a figure label}
\usage{
.addfiglab(lab, xl = par()$mar[2], yl = par()$mar[3], cex = 1)
}
\description{
Add a figure label based on x and y line values
}
