% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuncs.R
\name{calcMLCN}
\alias{calcMLCN}
\alias{.calcMLCN}
\title{Calculate the maximum likelihood copy number-state for a subject}
\usage{
.calcMLCN(N, mu, sf, phi, prior)
}
\arguments{
\item{N}{integer of length 1, the molecule counts for the subject}

\item{mu}{numeric of length 1, the locus mean counts for the group}

\item{sf}{numeric of length 1, the size-factor adjustment for the subject}

\item{phi}{numeric of length 1, the locus phi/dispersion value for the group}

\item{prior}{numeric of length 1, the prior probabilty of having a CNV}
}
\value{
list, first element gives the most likely copy number-state and
the second element gives the likelihood value
}
\description{
Calculate the maximum likelihood copy number-state for a subject
}
\details{
Need to add
}
