% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuncs.R
\name{callCN}
\alias{callCN}
\alias{.callCN}
\title{Call copy number-state}
\usage{
.callCN(cnts, min.dlt, max.its, prior)
}
\arguments{
\item{cnts}{data.table object containing the counts}

\item{min.dlt}{integer of length 1, the target number of changes in copy-
state to stop the alogorithm}

\item{max.its}{integer of length 1, the maximum number of iterations}

\item{prior}{numeric of length 1, the prior probability of having a CNV}
}
\details{
Need to add
}
