% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuncs.R
\name{clpsExon}
\alias{clpsExon}
\alias{.clpsExon}
\title{Collapse exons by specified width}
\usage{
.clpsExon(dat, cw)
}
\arguments{
\item{dat}{data.table object containing the counts}

\item{cw}{integer of length 1, the width to collapse by}
}
\description{
Collapse exons by specified width
}
\details{
Need to add
}
