% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvAggCall.R
\name{cnvAggCall}
\alias{cnvAggCall}
\title{Aggregate CN-state calls from different widths to single call}
\usage{
cnvAggCall(dat, width = 5, weight = FALSE)
}
\arguments{
\item{dat}{the input data, see details}

\item{width}{the maximum window size used to analyze the data}
}
\details{
Need to add
}
