% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{cnvBed2GRanges}
\alias{cnvBed2GRanges}
\alias{cnvBed2GRange}
\title{Read in .bed file to GRanges.}
\usage{
cnvBed2GRange(bedfile, pad = NULL, skip = 0)
}
\arguments{
\item{pad}{Integer of length 1, the number of basepairs to pad around the 
bed file intervals}

\item{intfile}{Character of length 1, the file path to the .interval file}
}
\value{
GRanges object with reference intervals
}
\description{
\code{cnvInt2GRanges} reads a GATK-style .interval file and 
returns a GRanges object for the intervals in the file.
}
\note{

}
