% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCallCN.R
\name{cnvCallCN}
\alias{cnvCallCN}
\title{Call copy number-state for the given data}
\usage{
cnvCallCN(
  counts,
  prior = 0.2,
  width = 1L,
  delta = 20L,
  iterations = 30L,
  verbose = TRUE
)
}
\arguments{
\item{counts}{data.table \link[=validObjects]{counts object}}

\item{prior}{numeric of length 1, the prior probability used to estimate
copy number variants; see details}

\item{width}{integer, window of contiguous intervals to calculate CN on; can
provide a vector with multiple widths (defaults to 1)}

\item{delta}{integer of length 1, minimum copy number state changes to stop
the algorithm; see details}

\item{iterations}{integer of length 1, the maximum number of iterations
before stopping algorithm; see details}

\item{verbose}{TRUE/FALSE}
}
\description{
Call copy number-state for the given data
}
\details{
Need to add
}
