% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCheckCor.R
\name{cnvCheckCor}
\alias{cnvCheckCor}
\title{Calculate sample-sample & sample-pool correlations}
\usage{
cnvCheckCor(counts)
}
\arguments{
\item{counts}{data.table \link[=validObjects]{counts object}}
}
\value{
List with two elements:
\describe{
\item{sampleSample}{pair-wise correlations}
\item{samplePool}{correlation of each sample to the sum counts
of all other samples}
}
}
\description{
Calculate sample-sample & sample-pool correlations
}
