% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCountsToMatrix.R
\name{cnvCountsToMatrix}
\alias{cnvCountsToMatrix}
\title{Convert counts object to interval by sample matrix}
\usage{
cnvCountsToMatrix(counts)
}
\arguments{
\item{counts}{data.table \link[=validObjects]{counts object}}
}
\value{
interval by sample count matrix
}
\description{
\code{cnvCountsToMatrix} is a convenience function for casting
a long-format \link[=validObjects]{counts object} into an interval by sample
counts matrix.
}
