% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGenPool.R
\name{cnvGenPool}
\alias{cnvGenPool}
\title{Generate a simulated pool of samples}
\usage{
cnvGenPool(ns, ne, wndw, sdlog = 0.688, pc = NULL, cs = NULL, cw,
  seed = NULL)
}
\arguments{
\item{ns}{integer of length 1, the number of samples}

\item{ne}{integer of length 1, the number of exons}

\item{wndw}{integer of length 2, the bounds for the number of molecules per
sample}

\item{sdlog}{numeric of length 1, the log sd for the log normal distribution
that defines the "genome" -- see details}

\item{pc}{numeric, the prior probability for the copy state}

\item{cs}{numeric, the possible copy states}

\item{cw}{integer of length 1, the width (number of exons) the cnv spans}

\item{seed}{numeric of length 1, the starting seed for the random number
generator, can be left NULL}

\item{pe}{numeric, the multinomial probability at each exon}

\item{seed.smpl}{logical, should the sample seed be given?}
}
\details{
The number of molecules per sample is drawn from a uniform distribution with
the bounds given by `wndw`.

For each pool a psuedo genome is created by modeling the probability of 
an exon receving a molecule during the capture process as a multinomial 
distribution, where each exon has its own probability. The multinomial 
distribution is given by drawing from a log-normal 
distribution with mean \code{log(1/ne)} and sd given by `sdlog`. The default 
sd was defined based on fitting the log normal distribution to a real pool 
of 16 samples and looking at the distribution of molecules per exon across 
the genome.

`cs`, `pc`, `cw` are all passed to \code{\link{cnvGenSmpl}}. Leaving `cs` &
`pc` NULL will use the default values. 

`seed` sets the seed for defining the multinomial distribution of 
probabilites that each exon gets a molecule. 

The number of reads per sample is passed as the seed to
\code{cnvGenSmpl}. This allows the simulated datasets to be completely
reproducible.
}
