% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGenSmpl.R
\name{cnvGenSmpl}
\alias{cnvGenSmpl}
\title{Generate a simulated sample}
\usage{
cnvGenSmpl(
  nr,
  pe,
  cw,
  seed = NULL,
  cs = c(0.001, 0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4),
  pc = c(1e-06, 0.000625, 0.998747, 0.000625, 1e-06, 2.5e-07, 2.5e-07, 2.5e-07,
    2.5e-07)
)
}
\arguments{
\item{nr}{integer of length 1, the number of molecules for the sample}

\item{pe}{numeric, the multinomial probability at each exon}

\item{cw}{integer of length 1, the width (number of exons) the cnv spans}

\item{seed}{numeric of length 1, the starting seed for the random number
generator, can be left NULL}

\item{cs}{numeric, the possible copy states}

\item{pc}{numeric, the prior probability for the copy state}
}
\description{
Generate a simulated sample
}
\details{
`pe` essentiall provides the "genome" and needs to be consistent across
samples that will be compared. 

`cs` and `pc` must be the same length. The possible copy states assume the 
normal copy state (2 copies) is 1. Likewise, a single copy of an exon would
be denoted as 0.5. The values we suggest for the possible states and their
priors are given as the default values.
}
