% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{cnvInt2GRanges}
\alias{cnvInt2GRanges}
\alias{cnvInt2GRange}
\title{Read in GATK-style .interval file to GRanges.}
\usage{
cnvInt2GRange(intfile, pad = NULL, skip = 0)
}
\arguments{
\item{intfile}{Character of length 1, the file path to the .interval file}

\item{pad}{Integer of length 1, the number of basepairs to pad around the 
bed file intervals}
}
\value{
GRanges object with reference intervals
}
\description{
\code{cnvInt2GRanges} reads a GATK-style .interval file and 
returns a GRanges object for the intervals in the file.
}
\note{

}
