% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{cnvReadRefs}
\alias{cnvReadRefs}
\title{Wrapper to Rsamtools:scanBamHeader to extract reference names & 
intervals from bam files.}
\usage{
cnvReadRefs(bamfile)
}
\arguments{
\item{bamfile}{Character of length 1, the file path to the .bam file}
}
\value{
IRanges object with reference intervals
}
\description{
\code{cnvReadRefs} serves as a wrapper to call 
Rsamtools:scanBamHeader and return IRanges object with reference intervals.
}
