% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvSimInterval.R
\name{cnvSimInterval}
\alias{cnvSimInterval}
\title{Generate a simulated interval}
\usage{
cnvSimInterval(
  n = 1e+05,
  seqnames = "simSeq",
  meanlog = -12.36,
  sdlog = 0.7393,
  seed = NULL
)
}
\arguments{
\item{n}{integer of length 1, the number of intervals}

\item{seqnames}{character of length 1, 'seqnames' for the simulated intervals}

\item{meanlog}{numeric of length 1, passed to \link[stats:rlnorm]{rlnorm}
to define distribution of interval capture probabilities}

\item{sdlog}{numeric of length 1, passed to \link[stats:rlnorm]{rlnorm}
to define distribution of interval capture probabilities}

\item{seed}{integer, passed to \link[base:set.seed]{set.seed()}}
}
\description{
Generate a simulated interval
}
\details{
\code{cnvSimInterval} generates an \link[=validObjects]{interval object} with
randomly-generated capture probabilities required for \code{\link[=cnvSimCounts]{cnvSimCounts()}}.
Interval capture probabilities are drawn initially from a lognormal
distribution, then normalized to a sum of 1.
The capture probabilities are stored in 'captureProb.'

The default values for meanlog & sdlog were derived from multiplexed whole-
exome capture of 16 human samples using the Agilent XT2 capture platform.

\code{cnvSimInterval} does not intend to make realistic exon sizes or
spacing.
All simulated intervals are 50 basepairs wide with start sites every 200
basepairs.
The structure of the object is simple and users can easily create their own
interval objects containing capture probabilities if greater complexity is
desired.

Setting the 'seed' parameter allows for reproducible simulation studies.
}
\seealso{
cnvSimCounts cnvSimPool
}
