% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvSimPool.R
\name{cnvSimPool}
\alias{cnvSimPool}
\title{Generate a simulated pool of samples}
\usage{
cnvSimPool(
  nSubjects = 16L,
  countRange = c(8000000L, 14000000L),
  interval = cnvSimInterval(),
  seed = NULL,
  ...
)
}
\arguments{
\item{nSubjects}{integer of length 1, the number of samples}

\item{countRange}{integer of length 2, the bounds for the number of
molecules per sample, see details}

\item{interval}{data.table \link[=validObjects]{interval object} with
'captureProb' field, see details}

\item{seed}{integer, passed to \link[base:set.seed]{set.seed()}}

\item{...}{passed to \code{\link[=cnvSimCounts]{cnvSimCounts()}}}
}
\description{
\code{cnvSimPool} is a convenience wrapper to generate a pool
of samples from the same interval object to mimic a multiplexed capture.
}
\details{
The number of molecules per sample is drawn from a uniform distribution with
the bounds given by 'countRange'.

The number of reads per sample is passed as the seed to \link{cnvSimCounts}.
This allows the simulated pools to be completely reproducible.
}
\seealso{
cnvSimCounts cnvSimPool
}
