% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterReads.R
\name{filterReads}
\alias{filterReads}
\alias{.filterReads}
\title{Modifies given data.table}
\usage{
.filterReads(x)
}
\arguments{
\item{x}{data.table object to be modified by reference}
}
\description{
\code{cnvFilterReads} modifies the given data.table object,
adding the 'ff' ("filter flag") field, specifying which reads to remove
from furhter analysis.
}
