% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{padGRanges}
\alias{padGRanges}
\alias{.padGRanges}
\title{Add padding to GRanges object}
\usage{
.padGRanges(g, pad)
}
\arguments{
\item{g}{GRanges object}

\item{pad}{Integer of length 1, the amount of padding to add}

\item{keyby}{Character, the field to key the data.table by}
}
\value{
GRanges object with padding added
}
\description{
\code{cnvReadBam} serves as a wrapper to call Rsamtools:scanBam
and processes the list result into a data.table.
}
