% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGetCounts.R
\name{procReads}
\alias{procReads}
\alias{.procReads}
\title{Process reads from a bam file}
\usage{
.procReads(rds, int, verbose)
}
\arguments{
\item{rds}{data.table with reads}

\item{int}{GRanges object with refs}

\item{verbose}{verbose}
}
\description{
This help file needs a lot of work
}
