% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{unpackQual}
\alias{unpackQual}
\alias{.unpackQual}
\title{Unpack the qual values from object returned by Rsamtools::scanBam}
\usage{
.unpackQual(l)
}
\arguments{
\item{l}{List object returned by Rsamtools::scanBam}
}
\value{
List object that resembles the output from Rsamtools:scanBam
}
\description{
\code{.unpackQual} converts the qual values to simple character
strings from the list returned by Rsamtools:scanBam. This allows for
collapsing the list into a data.table object.
}
