% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{unpackSeq}
\alias{unpackSeq}
\alias{.unpackSeq}
\title{Unpack the seq values from object returned by Rsamtools::scanBam}
\usage{
.unpackSeq(l)
}
\arguments{
\item{l}{List object returned by Rsamtools::scanBam}
}
\value{
List object that resembles the output from Rsamtools:scanBam
}
\description{
\code{.unpackSeq} converts the seq values to simple character 
strings from the list returned by Rsamtools:scanBam. This allows for 
collapsing the list into a data.table object.
}
