% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{unpackTag}
\alias{unpackTag}
\alias{.unpackTag}
\title{Unpack the tag values from object returned by Rsamtools::scanBam}
\usage{
.unpackTag(l, tags)
}
\arguments{
\item{l}{List object returned by Rsamtools::scanBam}

\item{tags}{Character, the tag names to unpack}
}
\value{
List object that resembles the output from Rsamtools:scanBam
}
\description{
\code{.unpackTag} simply extracts the tag values from the list
returned by Rsamtools:scanBam. This allows for collapsing the list into
a data.table object.
}
