% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityChecks.R
\name{validObjects}
\alias{validObjects}
\alias{cnvValidInterval}
\alias{cnvValidCounts}
\title{Functions to check for appropriate data structures}
\usage{
cnvValidInterval(x)

cnvValidCounts(x)
}
\description{
For efficiency, the mcCNV package does not currently formalize
data structures using the S4 system. Rather, the mcCNV package utilizes
\link[data.table:data.table]{data.table} objects, requiring specific fields.
}
\section{Interval objects}{

Interval objects define the intervals for computing copy number; inspired
by the GenomicRanges package, they require 'seqnames', 'start', and 'end'.
Additionally, the start and end fields must be integers and all end values
must be greater than all start values. Converting a
\link[GenomicRanges:GRanges]{GRanges} object with
\link[data.table:as.data.table]{as.data.table} will create a valid interval
object.

Of note, the \code{mcCNV} package uses 1-based positions (the standard for
R programming) for both the start and end positions.
\link[data.table:fread]{fread} can typically load BED files directly, but the
BED file specification uses 0-based start and 1-based end positions. Users
are responsible for ensuring the start and end positions are converted
appropriately, e.g. \code{interval[ , start := start + 1]}.
}

\section{Count objects}{

Count objects list the molecule counts over the given intervals. Count
objects require the same fields and specifications as interval objects
(listed above). Additionally, count objects have integer fields 'molCount'
giving the number of overlapping molecules & 'nCoverMult' giving the number
of the overlapping molecules that overlapped more than one interval.
Finally, count objects have a 'subject' field giving the subject represented.
Count objects are, by default, long and can be combined using the data.table
convention, \link[data.table:rbindlist]{rbindlist}. Count objects are created
by \link{cnvGetCounts}.

We provide the \link{cnvGatherCounts} convenience function for reading
\link[base:saveRDS]{saved count objects}, and combining them into a single
multiple-subject count object required by \link{cnvCallCN}.
}

