% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFuncs.R
\name{calcShrPhi}
\alias{calcShrPhi}
\alias{.calcShrPhi}
\title{Shrink phi values}
\usage{
.calcShrPhi(phi)
}
\arguments{
\item{phi}{numeric, the phi values to shrink}
}
\description{
Shrink phi values
}
\details{
Need to add.
}
