% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvAggCall.R
\name{cnvAggCall}
\alias{cnvAggCall}
\title{Aggregate CN-state calls from different widths to single call}
\usage{
cnvAggCall(calls)
}
\arguments{
\item{dat}{the input data, see details}
}
\description{
Aggregate CN-state calls from different widths to single call
}
\details{
Need to add

C1 -- TRUE if any interval of the given width that overlaps the exon was
called a CNV
}
