% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvAggCall.R
\name{cnvAggCall}
\alias{cnvAggCall}
\title{Aggregate CN-state calls from different widths to single call}
\usage{
cnvAggCall(dat, width = 5, weight = TRUE)
}
\arguments{
\item{dat}{the input data, see details}

\item{width}{integer of length 1, the maximum width (number of consecutive
exons) to evaluate simultaneously}

\item{weight}{logical of length 1, when TRUE likelihood values are multiplied
by the width}
}
\details{
Need to add
}
