% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCallCN.R
\name{cnvCallCN}
\alias{cnvCallCN}
\title{Call copy number-state for the given data}
\usage{
cnvCallCN(
  cnts,
  prior,
  width = 1:5,
  min.dlt = 20,
  max.its = 30,
  outfile = NULL,
  agg = FALSE,
  shrink = TRUE,
  keep.cols = NULL,
  verbose = FALSE,
  return.res = TRUE
)
}
\arguments{
\item{cnts}{the input data, see details}

\item{prior}{numeric of length 1, the prior probability of having a CNV}

\item{width}{integer, window of contiguous exons to calculate CN on; can 
provide a vector with multiple widths}

\item{min.dlt}{integer of length 1, the target number of changes in copy-
state to stop the alogorithm}

\item{max.its}{integer of length 1, the maximum number of iterations}

\item{outfile}{character of length 1, when not NULL, save .RDS file to the
file given}

\item{shrink}{logical of length 1, shrinkage applied to phi when TRUE}
}
\description{
Call copy number-state for the given data
}
\details{
Need to add
}
