% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCallCN.R
\name{cnvCallCN}
\alias{cnvCallCN}
\title{Call copy number-state for the given data}
\usage{
cnvCallCN(cnts, prior, width = 5, min.dlt = 20, max.its = 30,
  outfile = NULL, agg = FALSE, shrink = TRUE, keep.cols = NULL,
  verbose = FALSE)
}
\arguments{
\item{cnts}{the input data, see details}

\item{prior}{numeric of length 1, the prior probability of having a CNV}

\item{width}{integer of length 1, the maximum width (number of consecutive
exons) to evaluate simultaneously}

\item{min.dlt}{integer of length 1, the target number of changes in copy-
state to stop the alogorithm}

\item{max.its}{integer of length 1, the maximum number of iterations}

\item{outfile}{character of length 1, when not NULL, save .RDS file to the
file given}

\item{shrink}{logical of length 1, shrinkage applied to phi when TRUE}
}
\details{
Need to add
}
