% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvExomeDepth.R
\name{cnvExomeDepth}
\alias{cnvExomeDepth}
\title{Run ExomeDepth on mcCNV counts object}
\usage{
cnvExomeDepth(counts, transProb = 1e-04, cnvLength = 50000)
}
\arguments{
\item{counts}{data.table \link[=validObjects]{counts object}}

\item{transProb}{passed to \link[ExomeDepth:CallCNVs]{ExomeDepth::CallCNVs} 'transition.probability'}

\item{cnvLength}{passed to \link[ExomeDepth:CallCNVs]{ExomeDepth::CallCNVs} 'expected.CNV.length'}
}
\description{
Wrapper to run the ExomeDepth algorithm on a
\link[=validObjects]{counts object}
}
\details{
Runs ExomeDepth using the default parameters, then maps the call information
back to the counts object.
Will use 'mclapply' to parallelize the computation when available.
}
