% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGatherCounts.R
\name{cnvGatherCounts}
\alias{cnvGatherCounts}
\title{Gather saved count objects into one data.table}
\usage{
cnvGatherCounts(files, glob = TRUE, ...)
}
\arguments{
\item{files}{Character, file paths pointing to count objects; see details}

\item{glob}{Logical, should the given 'files' parameter be passed to
\link[base:Sys.glob]{Sys.glob()}? Defaults to TRUE}

\item{...}{arguments passed to \code{\link[parallel]{mclapply}}}
}
\description{
\code{cnvGatherCounts} reads \link[=validObjects]{count objects}
stored in disk space, and combines them into a single multiple-sample count
object.
}
\details{
The 'files' parameter must point to valid \link[=validObjects]{count objects}
saved to disk using the \link[base:saveRDS]{saveRDS()} function. When 'glob' is
\code{TRUE}, the default, any wildcards in 'files' will be expanded.
}
