% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGetCounts.R
\name{cnvGetCounts}
\alias{cnvGetCounts}
\title{Count molecules from a bam file overlapping given intervals}
\usage{
cnvGetCounts(bamfile, interval, subject = NULL, verbose = TRUE)
}
\arguments{
\item{bamfile}{Character of length 1, the file path to the .bam file}

\item{interval}{data.table object containing the intervals to count
molecules over; see \code{?cnvValidInterval}}

\item{subject}{Character of length 1, the subject name for bam file;
defaults to 'bamfile' parameter when NULL}

\item{verbose}{TRUE/FALSE}
}
\value{
data.table \link[=validObjects]{counts object}
}
\description{
\code{cnvGetCounts} takes a bam file and an
\link[=validObjects]{interval object} and returns the number of
overlapping molecules for each of the given intervals
}
\details{

}
