% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGetCounts.R
\name{cnvGetCounts}
\alias{cnvGetCounts}
\title{Given bam and int files, return molecule counts per interval}
\usage{
cnvGetCounts(bamfile, int, sbj = NULL, outfile = NULL,
  results = TRUE, verbose = FALSE)
}
\arguments{
\item{bamfile}{Character of length 1, the file path to the .bam file}

\item{int}{GRanges object containing the intervals (exons) to count reads
into}

\item{sbj}{Character or numeric of length 1, the subject name for bam file; 
defaults to the input filename when NULL}

\item{outfile}{Character of length 1, the file path to write results (.csv),
no results written if NULL}

\item{verbose}{TRUE/FALSE, if TRUE progress messages will be sent to the 
console}

\item{return}{TRUE/FALSE, should the results be returned to the console?
Should not be FALSE if outfile is NULL}
}
\value{
data.table object with counts for the intervals given in int
}
\description{
\code{cnvGetCounts} takes a bam file and an interval file and
returns the overlapping molecules for each interval in the interval file 
that span a given reference sequence.
}
