% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvGetCounts.R
\name{cnvGetCounts}
\alias{cnvGetCounts}
\title{Given bam and int files, return molecule counts per interval}
\usage{
cnvGetCounts(bamfile, intfile, refname, outfile = NULL, results = TRUE,
  verbose = FALSE)
}
\arguments{
\item{bamfile}{Character of length 1, the file path to the .bam file}

\item{intfile}{Character of length 1, the file path ot the interval file}

\item{refname}{Character of length 1, the name of the reference sequence
to evaluate}

\item{outfile}{Character of length 1, the file path to write results (.csv),
no results written if NULL}

\item{verbose}{TRUE/FALSE, if TRUE progress messages will be sent to the 
console}

\item{return}{TRUE/FALSE, should the results be returned to the console?
Should not be FALSE if outfile is NULL}
}
\value{
IRanges object with reference intervals
}
\description{
\code{cnvGetCounts} takes a bam file and an interval file and
returns the overlapping molecules for each interval in the interval file 
that span a given reference sequence.
}
