% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanBamUtils.R
\name{cnvReadBam}
\alias{cnvReadBam}
\title{Wrapper to Rsamtools:scanBam to load .bam file as a data.table object}
\usage{
cnvReadBam(
  bamfile,
  which,
  what = scanBamWhat(),
  flag = scanBamFlag(),
  tags = NULL,
  keyby = NULL
)
}
\arguments{
\item{bamfile}{Character of length 1, the file path to the .bam file}

\item{which}{IRangesList or GRanges object, the range of sequences to load,
passed to \link[Rsamtools:ScanBamParam]{Rsamtools::ScanBamParam} 'which' parameter; loads all sequences
when omitted}

\item{what}{Character, the fields to extract from the bam file, passed to
\link[Rsamtools:ScanBamParam]{Rsamtools::ScanBamParam} 'what' parameter}

\item{tags}{Character, the tag fields to extract from the bam file, passed to
\link[Rsamtools:ScanBamParam]{Rsamtools::ScanBamParam} 'tags' parameter}

\item{keyby}{Character, the field to key the data.table by}
}
\value{
data.table object with specified fields & tags extracted from
.bam file
}
\description{
\code{cnvReadBam} serves as a wrapper to call Rsamtools:scanBam
and processes the list result into a data.table.
}
